ttystatus/string.py
# Copyright 2010  Lars Wirzenius
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <http://www.gnu.org/licenses/>.


import ttystatus


class String(ttystatus.Widget):

    '''Display a value as a string.'''

    static_width = False

    def __init__(self, key):
        self._key = key
        self.value = ''

    def render(self, render):
        return str(self.value)

    def update(self, master):
        self.value = master[self._key]
